apps/core-api/pyproject.toml
[tool.poetry]
name = "core-api"
version = "0.1.0"
description = "Core API for Phi Agents"
authors = ["Phi Agents Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
openai = "^1.3.0"
pdfplumber = "^0.10.3"
python-docx = "^1.1.0"
pgvector = "^0.2.4"
pyyaml = "^6.0.1"
# Local package
phi-utils = {path = "../../packages/shared-utils", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

